+++ src/vs/editor/contrib/inlineCompletions/browser/utils.ts
@@ -6,10 +6,13 @@
 import { BugIndicatingError } from 'vs/base/common/errors';
 import { DisposableStore, IDisposable } from 'vs/base/common/lifecycle';
 import { IObservable, autorunOpts } from 'vs/base/common/observable';
+import { splitLines } from 'vs/base/common/strings';
 import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
+import { ISingleEditOperation } from 'vs/editor/common/core/editOperation';
 import { Position } from 'vs/editor/common/core/position';
 import { IRange, Range } from 'vs/editor/common/core/range';
 import { IModelDeltaDecoration } from 'vs/editor/common/model';
+import { TextModel } from 'vs/editor/common/model/textModel';
 
 export function applyEdits(text: string, edits: { range: IRange; text: string }[]): string {
 	const transformer = new PositionOffsetTransformer(text);
@@ -105,4 +108,60 @@ export function lengthOfText(text: string): Position {
 		}
 	}
 	return new Position(line, column);
+}
+
+/**
+ * Given some text edits, this function finds the new ranges of the editted text post application of all edits.
+ * Assumes that the edit ranges are disjoint
+ * @param edits edits applied
+ * @returns new ranges post edits for every edit
+ */
+export function getNewRanges(edits: ISingleEditOperation[]): Range[] {
+
+	if (edits.length === 0) {
+		return [];
+	}
+
+	const sortIndices = Array.from(edits.keys()).sort((a, b) =>
+		Range.compareRangesUsingStarts(edits[a].range, edits[b].range)
+	);
+	const ranges: Range[] = [];
+	let previousEditEndLineNumber = 0;
+	let positionOffset = new Position(0, 0);
+
+	for (const index of sortIndices) {
+		const edit = edits[index];
+		const text = edit.text ?? '';
+		const rangeStart = Position.lift({
+			lineNumber: edit.range.startLineNumber + positionOffset.lineNumber,
+			column: edit.range.startColumn + (edit.range.startLineNumber === previousEditEndLineNumber ? positionOffset.column : 0)
+		});
+		const rangeEnd = addPositions(
+			rangeStart,
+			lengthOfText(text)
+		);
+		ranges.push(Range.fromPositions(rangeStart, rangeEnd));
+		const splitText = splitLines(text);
+		previousEditEndLineNumber = edit.range.endLineNumber;
+		positionOffset = positionOffset.delta(
+			splitText.length - edit.range.endLineNumber + edit.range.startLineNumber - 1,
+			rangeEnd.column - edit.range.endColumn - positionOffset.column
+		);
+	}
+	return ranges.map((_, index) => ranges[sortIndices.indexOf(index)]);
+}
+
+/**
+ * Given a text model and edits, this function finds the inverse text edits
+ * @param model model on which to apply the edits
+ * @param edits edits applied
+ * @returns inverse edits
+ */
+export function inverseEdits(model: TextModel, edits: ISingleEditOperation[]): ISingleEditOperation[] {
+	const newRanges = getNewRanges(edits);
+	const inverseEdits: ISingleEditOperation[] = [];
+	for (let i = 0; i < edits.length; i++) {
+		inverseEdits.push({ range: newRanges[i], text: model.getValueInRange(edits[i].range) });
+	}
+	return inverseEdits;
 }
