+++ src/vs/editor/test/common/model/bracketPairColorizer/combineTextEditInfos.test.ts
@@ -58,7 +58,7 @@ function runTest(seed: number) {
 	textModelS2.dispose();
 }
 
-function getRandomEditInfos(textModel: TextModel, count: number, rng: MersenneTwister): TextEditInfo[] {
+export function getRandomEditInfos(textModel: TextModel, count: number, rng: MersenneTwister): TextEditInfo[] {
 	const edits: TextEditInfo[] = [];
 	let i = 0;
 	for (let j = 0; j < count; j++) {
@@ -79,7 +79,7 @@ function getRandomEdit(textModel: TextModel, rangeOffsetStart: number, rng: Mers
 	return new TextEditInfo(positionToLength(textModel.getPositionAt(offsetStart)), positionToLength(textModel.getPositionAt(offsetEnd)), toLength(lineCount, columnCount));
 }
 
-function toEdit(editInfo: TextEditInfo): ISingleEditOperation {
+export function toEdit(editInfo: TextEditInfo): ISingleEditOperation {
 	const l = lengthToObj(editInfo.newLength);
 	let text = '';
 
@@ -100,7 +100,7 @@ function toEdit(editInfo: TextEditInfo): ISingleEditOperation {
 }
 
 // Generated by copilot
-class MersenneTwister {
+export class MersenneTwister {
 	private readonly mt = new Array(624);
 	private index = 0;
